+++ chain/chain/src/state_sync/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) use adapter::ChainStateSyncAdapter;
+pub(crate) use utils::{is_sync_prev_hash, update_sync_hashes};
+
+mod adapter;
+mod state_request_tracker;
+mod utils;
