+++ chain/client/src/debug.rs
@@ -186,7 +186,7 @@ impl Handler<DebugStatus> for ClientActorInner {
                 Ok(DebugStatusResponse::CatchupStatus(self.client.get_catchup_status()?))
             }
             DebugStatus::RequestedStateParts => Ok(DebugStatusResponse::RequestedStateParts(
-                self.client.chain.get_requested_state_parts(),
+                self.client.chain.state_sync_adapter.get_requested_state_parts(),
             )),
             DebugStatus::ChainProcessingStatus => Ok(DebugStatusResponse::ChainProcessingStatus(
                 self.client.chain.get_chain_processing_info(),
