+++ chain/client/src/stateless_validation/shadow_validate.rs
@@ -1,11 +1,11 @@
 use crate::stateless_validation::state_witness_producer::CreateWitnessResult;
+use crate::Client;
+use near_chain::get_chunk_clone_from_header;
 use near_chain::stateless_validation::chunk_validation::validate_prepared_transactions;
 use near_chain::types::{RuntimeStorageConfig, StorageDataSource};
 use near_chain::{Block, BlockHeader};
 use near_chain_primitives::Error;
 use near_primitives::sharding::{ShardChunk, ShardChunkHeader};
-
-use crate::Client;
 
 impl Client {
     // Temporary feature to make node produce state witness for every chunk in every processed block
@@ -24,7 +24,7 @@ impl Client {
             .enumerate()
             .filter(|(_, chunk)| chunk.is_new_chunk(block.header().height()))
         {
-            let chunk = self.chain.get_chunk_clone_from_header(chunk)?;
+            let chunk = get_chunk_clone_from_header(&self.chain.chain_store, chunk)?;
             // TODO(resharding) This doesn't work if shard layout changes.
             let prev_chunk_header = prev_block_chunks.get(shard_index).unwrap();
             if let Err(err) =
