+++ chain/client/src/view_client_actor.rs
@@ -1356,13 +1356,15 @@ impl Handler<StateRequestHeader> for ViewClientActorInner {
             return None;
         }
         let header = match self.chain.check_sync_hash_validity(&sync_hash) {
-            Ok(true) => match self.chain.get_state_response_header(shard_id, sync_hash) {
-                Ok(header) => Some(header),
-                Err(err) => {
-                    error!(target: "sync", ?err, "Cannot build state sync header");
-                    None
+            Ok(true) => {
+                match self.chain.state_sync_adapter.get_state_response_header(shard_id, sync_hash) {
+                    Ok(header) => Some(header),
+                    Err(err) => {
+                        error!(target: "sync", ?err, "Cannot build state sync header");
+                        None
+                    }
                 }
-            },
+            }
             Ok(false) => {
                 warn!(target: "sync", ?sync_hash, "sync_hash didn't pass validation, possible malicious behavior");
                 // Don't respond to the node, because the request is malformed.
@@ -1432,7 +1434,11 @@ impl Handler<StateRequestPart> for ViewClientActorInner {
         tracing::debug!(target: "sync", ?shard_id, ?sync_hash, ?part_id, "Computing state request part");
         let part = match self.chain.check_sync_hash_validity(&sync_hash) {
             Ok(true) => {
-                let part = match self.chain.get_state_response_part(shard_id, part_id, sync_hash) {
+                let part = match self
+                    .chain
+                    .state_sync_adapter
+                    .get_state_response_part(shard_id, part_id, sync_hash)
+                {
                     Ok(part) => Some((part_id, part)),
                     Err(err) => {
                         error!(target: "sync", ?err, ?sync_hash, ?shard_id, part_id, "Cannot build state part");
