+++ chain/epoch-manager/src/shard_tracker.rs
@@ -4,8 +4,10 @@ use crate::EpochManagerAdapter;
 use itertools::Itertools;
 use near_cache::SyncLruCache;
 use near_chain_configs::ClientConfig;
+use near_chain_primitives::Error;
 use near_primitives::errors::EpochError;
 use near_primitives::hash::CryptoHash;
+use near_primitives::sharding::StateSyncInfo;
 use near_primitives::types::{AccountId, EpochId, ShardId};
 
 #[derive(Clone)]
@@ -277,6 +279,81 @@ impl ShardTracker {
     /// Returns whether the node is configured for all shards tracking.
     pub fn tracks_all_shards(&self) -> bool {
         matches!(self.tracked_config, TrackedConfig::AllShards)
+    }
+
+    /// Return all shards that whose states need to be caught up
+    /// That has two cases:
+    /// 1) Shard layout will change in the next epoch. In this case, the method returns all shards
+    ///    in the current epoch that will be split into a future shard that `me` will track.
+    /// 2) Shard layout will be the same. In this case, the method returns all shards that `me` will
+    ///    track in the next epoch but not this epoch
+    fn get_shards_to_state_sync(
+        &self,
+        me: &Option<AccountId>,
+        parent_hash: &CryptoHash,
+    ) -> Result<Vec<ShardId>, Error> {
+        let epoch_id = self.epoch_manager.get_epoch_id_from_prev_block(parent_hash)?;
+        let mut shards_to_sync = Vec::new();
+        for shard_id in self.epoch_manager.shard_ids(&epoch_id)? {
+            if self.should_catch_up_shard(me, parent_hash, shard_id)? {
+                shards_to_sync.push(shard_id)
+            }
+        }
+        Ok(shards_to_sync)
+    }
+
+    /// Returns whether we need to initiate state sync for the given `shard_id` for the epoch
+    /// beginning after the block `epoch_last_block`. If that epoch is epoch T, the logic is:
+    /// - will track the shard in epoch T+1
+    /// - AND not tracking it in T
+    /// - AND didn't track it in T-1
+    /// We check that we didn't track it in T-1 because if so, and we're in the relatively rare case
+    /// where we'll go from tracking it to not tracking it and back to tracking it in consecutive epochs,
+    /// then we can just continue to apply chunks as if we were tracking it in epoch T, and there's no need to state sync.
+    fn should_catch_up_shard(
+        &self,
+        me: &Option<AccountId>,
+        prev_hash: &CryptoHash,
+        shard_id: ShardId,
+    ) -> Result<bool, Error> {
+        // Won't care about it next epoch, no need to state sync it.
+        if !self.will_care_about_shard(me.as_ref(), prev_hash, shard_id, true) {
+            return Ok(false);
+        }
+        // Currently tracking the shard, so no need to state sync it.
+        if self.care_about_shard(me.as_ref(), prev_hash, shard_id, true) {
+            return Ok(false);
+        }
+
+        // Now we need to state sync it unless we were tracking the parent in the previous epoch,
+        // in which case we don't need to because we already have the state, and can just continue applying chunks
+
+        let tracked_before =
+            self.cared_about_shard_in_prev_epoch(me.as_ref(), prev_hash, shard_id, true);
+        Ok(!tracked_before)
+    }
+
+    /// Return a StateSyncInfo that includes the information needed for syncing state for shards needed
+    /// in the next epoch.
+    pub fn get_state_sync_info(
+        &self,
+        me: &Option<AccountId>,
+        epoch_id: &EpochId,
+        block_hash: &CryptoHash,
+        prev_hash: &CryptoHash,
+    ) -> Result<Option<StateSyncInfo>, Error> {
+        let shards_to_state_sync = self.get_shards_to_state_sync(me, prev_hash)?;
+        if shards_to_state_sync.is_empty() {
+            Ok(None)
+        } else {
+            tracing::debug!(target: "chain", "Downloading state for {:?}, I'm {:?}", shards_to_state_sync, me);
+            let protocol_version = self.epoch_manager.get_epoch_protocol_version(epoch_id)?;
+            // Note that this block is the first block in an epoch because this function is only called
+            // in get_catchup_and_state_sync_infos() when that is the case.
+            let state_sync_info =
+                StateSyncInfo::new(protocol_version, *block_hash, shards_to_state_sync);
+            Ok(Some(state_sync_info))
+        }
     }
 }
