+++ integration-tests/src/test_loop/tests/bandwidth_scheduler.rs
@@ -23,7 +23,7 @@ use near_async::test_loop::futures::TestLoopFutureSpawner;
 use near_async::test_loop::sender::TestLoopSender;
 use near_async::test_loop::TestLoopV2;
 use near_async::time::Duration;
-use near_chain::{ChainStoreAccess, ReceiptFilter};
+use near_chain::{get_incoming_receipts_for_shard, ChainStoreAccess, ReceiptFilter};
 use near_chain_configs::test_genesis::{
     build_genesis_and_epoch_config_store, GenesisAndEpochConfigParams, ValidatorsSpec,
 };
@@ -304,17 +304,16 @@ fn analyze_workload_blocks(
             cur_chunk_stats.congestion_level = congestion_control.congestion_level();
 
             // Fetch incoming receipts for this chunk
-            let incoming_receipts_proofs = chain
-                .chain_store
-                .get_incoming_receipts_for_shard(
-                    epoch_manager,
-                    shard_id,
-                    &cur_shard_layout,
-                    *block.hash(),
-                    prev_height_included,
-                    ReceiptFilter::TargetShard,
-                )
-                .unwrap();
+            let incoming_receipts_proofs = get_incoming_receipts_for_shard(
+                &chain.chain_store,
+                epoch_manager,
+                shard_id,
+                &cur_shard_layout,
+                *block.hash(),
+                prev_height_included,
+                ReceiptFilter::TargetShard,
+            )
+            .unwrap();
 
             // Fetch outgoing receipts generated by this chunk
             let outgoing_receipts =
