+++ integration-tests/src/tests/client/sync_state_nodes.rs
@@ -560,15 +560,22 @@ fn ultra_slow_test_dump_epoch_missing_chunk_in_last_block() {
             let sync_prev_block = &blocks[sync_block_idx - 1];
             let sync_prev_height_included = sync_prev_block.chunks()[0].height_included();
 
-            let state_sync_header =
-                env.clients[0].chain.get_state_response_header(shard_id, sync_hash).unwrap();
+            let state_sync_header = env.clients[0]
+                .chain
+                .state_sync_adapter
+                .get_state_response_header(shard_id, sync_hash)
+                .unwrap();
             let num_parts = state_sync_header.num_state_parts();
             let state_root = state_sync_header.chunk_prev_state_root();
             // Check that state parts can be obtained.
             let state_sync_parts: Vec<_> = (0..num_parts)
                 .map(|i| {
                     // This should obviously not fail, aka succeed.
-                    env.clients[0].chain.get_state_response_part(shard_id, i, sync_hash).unwrap()
+                    env.clients[0]
+                        .chain
+                        .state_sync_adapter
+                        .get_state_response_part(shard_id, i, sync_hash)
+                        .unwrap()
                 })
                 .collect();
 
@@ -591,11 +598,13 @@ fn ultra_slow_test_dump_epoch_missing_chunk_in_last_block() {
             tracing::info!(target: "test", "state sync - set parts");
             env.clients[1]
                 .chain
+                .state_sync_adapter
                 .set_state_header(shard_id, sync_hash, state_sync_header.clone())
                 .unwrap();
             for i in 0..num_parts {
                 env.clients[1]
                     .chain
+                    .state_sync_adapter
                     .set_state_part(
                         shard_id,
                         sync_hash,
